Fix keyboard depress wait
key_state contains garbage the first time it is passed to portG_read, which in
turn checks if the pin is in input mode. If it isn't, it returns -1, leaving
key_state in an undefined state. Therefore, the check key_state == LOW will
almost always pass.

This does not work if we're still pressing the key when this check happens, it
will pass and clear the interrupt even if we are still pressing the key.

To fix it, initialize key_state to LOW, and on portG_read, only abort if the
mode is not input or EINT. Even if the documentation says that reading from a
port while in EINT mode is undefined, in our particular case, it works.